pyproject.toml
[project]
name = "deriv-translation-pipeline"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.43",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "litellm>=1.40",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "pillow>=10.0",
]

[project.scripts]
translate = "translate:main"
